Removed log4j and topcoder loggers
The code base was using (at least) three distinct logging tools before
this commit.  As discussed in #16 we're going to just use the built in
logging for now, since none of the features provided by log4j are
compelling enough to warrant a logging dependency in this project.

Some of the Top coder logging code used log/Log instead of
logger/Logger -- in addition to changing the log implementations I
also attempted to make the variable / method names consistant to use
Logger.